requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
gigachat==0.1.42.post2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
numpy==2.3.3
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.0.0
python-telegram-bot==20.3
RapidFuzz==3.14.1
requests==2.31.0
sniffio==1.3.1
thefuzz==0.22.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0